+++ packages/react/src/index.ts
@@ -240,6 +240,9 @@ export * from './components/primitives/DatePicker/DatePicker';
 export {default as Checkbox} from './components/primitives/Checkbox/Checkbox';
 export * from './components/primitives/Checkbox/Checkbox';
 
+export {default as Toggle} from './components/primitives/Toggle/Toggle';
+export * from './components/primitives/Toggle/Toggle';
+
 export {default as FormControl} from './components/primitives/FormControl/FormControl';
 export * from './components/primitives/FormControl/FormControl';
 
@@ -289,6 +292,15 @@ export {
   http,
   getActiveTheme,
   navigate,
+  resolveEmojiUrisInHtml,
+  isEmojiUri,
+  EMOJI_URI_SCHEME,
+  extractEmojiFromUri,
+  resolveMeta,
+  resolveFlowTemplateLiterals,
+  countryCodeToFlagEmoji,
+  resolveLocaleDisplayName,
+  resolveLocaleEmoji,
   // Export `v2` models and types as first class citizens since they are
   // going to be the primary way to interact with embedded flows moving forward.
   EmbeddedFlowComponentTypeV2 as EmbeddedFlowComponentType,