+++ packages/browser/CHANGELOG.md
@@ -1,5 +1,54 @@
 # @asgardeo/browser
 
+## 0.5.9
+
+### Patch Changes
+
+- [#426](https://github.com/asgardeo/javascript/pull/426)
+  [`29463c4`](https://github.com/asgardeo/javascript/commit/29463c4be6ba7bd612ada83ccbe8179a2cd374f6) Thanks
+  [@kavindadimuthu](https://github.com/kavindadimuthu)! - Remove esbuild-plugin-polyfill-node &
+  esbuild-plugins-node-modules-polyfill dependencies
+
+## 0.5.8
+
+### Patch Changes
+
+- Updated dependencies
+  [[`f53ef78`](https://github.com/asgardeo/javascript/commit/f53ef78da1fc6972675eef013df5f53f39a0a764)]:
+  - @asgardeo/javascript@0.12.0
+
+## 0.5.7
+
+### Patch Changes
+
+- Updated dependencies
+  [[`db123a8`](https://github.com/asgardeo/javascript/commit/db123a85e6dcd4406d61cb67972269412eccb92d)]:
+  - @asgardeo/javascript@0.11.0
+
+## 0.5.6
+
+### Patch Changes
+
+- Updated dependencies
+  [[`f380a3a`](https://github.com/asgardeo/javascript/commit/f380a3a137d2a08a921373c5a9ca1ac4d588ad45)]:
+  - @asgardeo/javascript@0.10.2
+
+## 0.5.5
+
+### Patch Changes
+
+- Updated dependencies
+  [[`0dca4b4`](https://github.com/asgardeo/javascript/commit/0dca4b405afe75d6e5baf0a2de2f3f204780dcf9)]:
+  - @asgardeo/javascript@0.10.1
+
+## 0.5.4
+
+### Patch Changes
+
+- [#409](https://github.com/asgardeo/javascript/pull/409)
+  [`0adb9bc`](https://github.com/asgardeo/javascript/commit/0adb9bc633d932802b2829c79e7aa6c06229daf3) Thanks
+  [@brionmario](https://github.com/brionmario)! - Fix issues in Language Dropdown & expose emoji resolving utils
+
 ## 0.5.3
 
 ### Patch Changes